lift load previews into zustand
